Make sure MakeMC doesn't fail silently when evtgen config files cannot be copied over

+++ MakeMC.sh
@@ -1275,15 +1275,27 @@ if [[ "$GENERATOR_POST" != "No" ]]; then
 	echo $GENERATOR_POST_CONFIGDEC
 	if [[ "$GENERATOR_POST_CONFIG" != "default" ]]; then
 		cp $GENERATOR_POST_CONFIG ./post'_'$GENERATOR_POST'_'$formatted_runNumber'_'$formatted_fileNumber.conf
+    if [[ ! -f ./post'_'$GENERATOR_POST'_'$formatted_runNumber'_'$formatted_fileNumber.conf ]]; then
+      echo "Couldn't copy $GENERATOR_POST_CONFIG. Exit."
+      exit 1
+    fi
 	fi
 
 	if [[ "$GENERATOR_POST" == "decay_evtgen" ]]; then
 		if [[ "$GENERATOR_POST_CONFIGEVT" != "default" ]]; then
       cp $GENERATOR_POST_CONFIGEVT ./postevt'_'$GENERATOR_POST'_'$formatted_runNumber'_'$formatted_fileNumber.conf
+      if [[ ! -f ./postevt'_'$GENERATOR_POST'_'$formatted_runNumber'_'$formatted_fileNumber.conf ]]; then
+        echo "Couldn't copy $GENERATOR_POST_CONFIGEVT. Exit."
+        exit 1
+      fi
 			export EVTGEN_PARTICLE_DEFINITIONS=$PWD/postevt'_'$GENERATOR_POST'_'$formatted_runNumber'_'$formatted_fileNumber.conf
 		fi
 		if [[ "$GENERATOR_POST_CONFIGDEC" != "default" ]];then
       cp $GENERATOR_POST_CONFIGDEC ./postdec'_'$GENERATOR_POST'_'$formatted_runNumber'_'$formatted_fileNumber.conf
+      if [[ ! -f ./postdec'_'$GENERATOR_POST'_'$formatted_runNumber'_'$formatted_fileNumber.conf ]]; then
+        echo "Couldn't copy $GENERATOR_POST_CONFIGDEC. Exit."
+        exit 1
+      fi
 			export EVTGEN_DECAY_FILE=$PWD/postdec'_'$GENERATOR_POST'_'$formatted_runNumber'_'$formatted_fileNumber.conf
 		fi
 		echo decay_evtgen -o$STANDARD_NAME'_decay_evtgen'.hddm -upost'_'$GENERATOR_POST'_'$formatted_runNumber'_'$formatted_fileNumber.conf $STANDARD_NAME.hddm
